Move `.Invoke()` and `.Assign()` to `ObjectExtensions.cs`
